requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
Faker==19.13.0
plotly==5.14.1
pandas==2.0.3
numpy==1.24.3
mapbox==0.18.0 